requirements.txt
Django
django-redis
django-rest-framework
djoser
django-cors-headers
pandas
xlrd
